+++ frontend/components/index.ts
@@ -15,6 +15,7 @@ export * from './ProfileNav'
 export * from './ProfilePicture'
 export * from './RecipeDuration'
 export * from './RecipeList'
+export * from './RecipeNav'
 export * from './RecipePreview'
 export * from './RecipeTitle'
 export * from './RecipeVersion'
