+++ frontend/pages/recipe.tsx
@@ -1,43 +1,21 @@
-import React, { useState, useContext } from 'react'
-import Router from 'next/router'
+import React from 'react'
 import * as _ from 'lodash'
-import {
-  Alert,
-  FormGroup,
-  Row,
-  Col,
-  Label,
-  Input,
-  Modal,
-  ModalBody,
-  ButtonDropdown,
-  DropdownToggle,
-  DropdownMenu,
-  DropdownItem,
-  Button
-} from 'reactstrap'
+
 import {
   Head,
   Layout,
-  IfLoggedIn,
   RecipeVersion as RecipeVersionView,
-  RecipeTitle
+  RecipeTitle,
+  RecipeNav
 } from '../components'
 import { RecipeJSON } from '../models/recipe'
 import { RecipeVersionJSON } from '../models/recipe_version'
-import {
-  PlateZeroApiError,
-  getRecipe,
-  getRecipeVersion,
-  deleteRecipe,
-  patchRecipe
-} from '../common/http'
-import { UserContext } from '../context/UserContext'
-import { TokenContext } from '../context/TokenContext'
+import { getRecipe, getRecipeVersion } from '../common/http'
 
 interface Props {
   recipe: RecipeJSON
   recipeVersion?: RecipeVersionJSON
+  pathname: string
 }
 
 interface State {
@@ -45,7 +23,7 @@ interface State {
 }
 
 export default class Recipe extends React.Component<Props, State> {
-  static async getInitialProps({ query }): Promise<Props> {
+  static async getInitialProps({ pathname, query }): Promise<Props> {
     const recipe = await getRecipe(query.username, query.slug)
     const masterBranch = _.head(
       _.filter(recipe.branches, r => r.name === 'master')
@@ -54,7 +32,7 @@ export default class Recipe extends React.Component<Props, State> {
     const recipeVersion = versionId
       ? await getRecipeVersion(query.username, query.slug, versionId)
       : undefined
-    return { recipe, recipeVersion }
+    return { recipe, recipeVersion, pathname }
   }
 
   constructor(props) {
@@ -77,183 +55,12 @@ export default class Recipe extends React.Component<Props, State> {
           image={recipe.image_url}
           url={`/${recipe.owner.username}/${recipe.slug}`}
         />
-        <Row className="mt-3">
-          <Col>
-            <RecipeTitle recipe={recipe} />
-          </Col>
-          <Col xs="auto">
-            <IfLoggedIn username={recipe.owner.username}>
-              <ActionMenu recipe={recipe} />
-            </IfLoggedIn>
-          </Col>
-        </Row>
+        <div className="mt-3">
+          <RecipeTitle recipe={recipe} />
+        </div>
+        <RecipeNav recipe={recipe} route={this.props.pathname} />
         {recipeVersion && <RecipeVersionView recipeVersion={recipeVersion} />}
       </Layout>
     )
   }
-}
-
-const DeleteModal = ({
-  recipe,
-  isOpen,
-  toggle,
-  close
-}: {
-  recipe: RecipeJSON
-  isOpen: boolean
-  toggle: () => void
-  close: () => void
-}) => {
-  const user = useContext(UserContext)
-  const token = useContext(TokenContext)
-  const handleDelete = async () => {
-    try {
-      await deleteRecipe(recipe.slug, { token })
-    } catch {}
-    Router.push(`/${user.username}`)
-  }
-  return (
-    <Modal isOpen={isOpen} toggle={toggle}>
-      <ModalBody>
-        <h5>Permanently delete {recipe.title}</h5>
-        <p>
-          Are you sure you want to permanently delete{' '}
-          <strong>{recipe.title}</strong>?
-        </p>
-        <Button color="danger" block onClick={handleDelete}>
-          Yes, delete it forever
-        </Button>
-        <Button
-          color="link"
-          className="text-muted"
-          outline
-          block
-          onClick={close}
-        >
-          Never mind, keep it for now
-        </Button>
-      </ModalBody>
-    </Modal>
-  )
-}
-
-const RenameModal = ({
-  recipe,
-  isOpen,
-  toggle,
-  close
-}: {
-  recipe: RecipeJSON
-  isOpen: boolean
-  toggle: () => void
-  close: () => void
-}) => {
-  const [title, setTitle] = useState(recipe.title)
-  const [subtitle, setSubtitle] = useState(recipe.subtitle || '')
-  const [description, setDescription] = useState(recipe.description || '')
-  const [errors, setErrors] = useState([])
-  const token = useContext(TokenContext)
-  const handleSave = async () => {
-    const patch = { title, subtitle, description }
-    setErrors([])
-    try {
-      await patchRecipe(recipe.slug, patch, { token })
-      Router.push(`/${recipe.owner.username}/${recipe.slug}`)
-    } catch (e) {
-      if (e instanceof PlateZeroApiError) {
-        setErrors(e.messages)
-      } else {
-        setErrors([e])
-      }
-    }
-  }
-  return (
-    <Modal isOpen={isOpen} toggle={toggle} size="lg">
-      <ModalBody>
-        <FormGroup>
-          <Label>Recipe Title</Label>
-          <Input value={title} onChange={e => setTitle(e.target.value)} />
-        </FormGroup>
-        <FormGroup>
-          <Label>Subtitle</Label>
-          <Input value={subtitle} onChange={e => setSubtitle(e.target.value)} />
-        </FormGroup>
-        <FormGroup>
-          <Label>Description</Label>
-          <Input
-            type="textarea"
-            value={description}
-            onChange={e => setDescription(e.target.value)}
-          />
-        </FormGroup>
-        {errors.map((e, key) => (
-          <Alert color="danger" key={key}>
-            {e}
-          </Alert>
-        ))}
-        <Button color="success" block onClick={handleSave}>
-          Save
-        </Button>
-        <Button
-          color="link"
-          className="text-muted"
-          outline
-          block
-          onClick={close}
-        >
-          Never mind
-        </Button>
-      </ModalBody>
-    </Modal>
-  )
-}
-
-const ActionMenu = ({ recipe }: { recipe: RecipeJSON }) => {
-  const [isOpen, setOpen] = useState(false)
-  const [isRenameModalOpen, setRenameModalOpen] = useState(false)
-  const [isDeleteModalOpen, setDeleteModalOpen] = useState(false)
-  return (
-    <>
-      <ButtonDropdown isOpen={isOpen} toggle={() => setOpen(!isOpen)}>
-        <DropdownToggle
-          caret
-          outline
-          color="primary"
-          onClick={() => setOpen(true)}
-        >
-          Actions
-        </DropdownToggle>
-        <DropdownMenu right>
-          <DropdownItem onClick={() => setRenameModalOpen(true)}>
-            Edit Title and Description
-          </DropdownItem>
-          <DropdownItem
-            href={`/${recipe.owner.username}/${
-              recipe.slug
-            }/branches/master/edit`}
-          >
-            Update Recipe
-          </DropdownItem>
-          <DropdownItem
-            onClick={() => setDeleteModalOpen(true)}
-            className="text-danger"
-          >
-            Delete Recipe&hellip;
-          </DropdownItem>
-        </DropdownMenu>
-      </ButtonDropdown>
-      <RenameModal
-        isOpen={isRenameModalOpen}
-        recipe={recipe}
-        toggle={() => setRenameModalOpen(!isRenameModalOpen)}
-        close={() => setRenameModalOpen(false)}
-      />
-      <DeleteModal
-        isOpen={isDeleteModalOpen}
-        recipe={recipe}
-        toggle={() => setDeleteModalOpen(!isDeleteModalOpen)}
-        close={() => setDeleteModalOpen(false)}
-      />
-    </>
-  )
 }
